Notion of ElementData
git-svn-id: https://geuz.org/svn/gmsh/trunk@14770 59c8cabf-10d6-4a28-b43f-3936b85956c4

+++ projects/small_fem/assembler/Dof.cpp
@@ -12,7 +12,7 @@ Dof::Dof(const Dof& other){
   this->type    = other.type;
 }
 
-Dof::Dof(unsigned int entity, unsigned int type){
+Dof::Dof(unsigned long int entity, unsigned int type){
   this->entity  = entity;
   this->type    = type;
 }
@@ -22,7 +22,7 @@ Dof::~Dof(void){
 
 std::string Dof::toString(void) const{
   std::stringstream s;
-  
+
   s << "(" << entity << ", " << type << ")";
 
   return s.str();
